Code review changes * Rename ConceptPath.tsx to ConceptPathDisplay.tsx
