+++ packages/components/src/internal/components/ontology/ConceptPathDisplay.tsx
